Fix auto-detection of radeon device
If the Bus ID is not given, do not assume that /dev/dri/card0 is the radeon
device. This is not the case on this desktop PC with integrated Intel graphics
and an AMD GPU.

Note that `drmOpen("radeon", NULL)` will fail when it is in use, that is why the
Bus ID is explicitly given here.