Add optional unprivileged mode in Xorg
When running radeontop in a X session, the X server is the DRM Master
and must authorize our token. Add the required logic in a shared
library such that distributions can mark libxcb as optional dependency.
This allows radeontop to be installed without pulling in a full X stack
on a headless OpenCL server.