Add support for UVD and VCE status
This add a bar for UVD and VCE.

Tested on RX 460 (POLARIS).

This was a difficult and tedious port of changes whose underlying code had
changed extensively over the 5 intervening years.
The original is at https://github.com/clbr/radeontop/pull/45 , but this
port did not contact the author.

Fixes #29. Tested on CARRIZO, STONEY and KAVERI.

Added GTT memory usage details

Co-authored-by: Chandu Babu Namburu <chandu@amd.com>